feat: add limit and offset to retrieveAllMovies and chnaged price type to NUMERIC

+++ sql/create_table.sql
@@ -3,5 +3,5 @@ CREATE TABLE movies (
   name VARCHAR(50) UNIQUE NOT NULL,
   category VARCHAR(20) NOT NULL,
   duration INTEGER NOT NULL,
-  price INTEGER NOT NULL
+  price NUMERIC(10,2) NOT NULL
 );
